+++ template_variables/src/epilot_template_variables/sdkconfiguration.py
@@ -26,10 +26,10 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "1.0.0"
-    sdk_version: str = "1.6.0"
-    gen_version: str = "2.415.6"
+    sdk_version: str = "1.6.1"
+    gen_version: str = "2.415.7"
     user_agent: str = (
-        "speakeasy-sdk/python 1.6.0 2.415.6 1.0.0 epilot-template-variables"
+        "speakeasy-sdk/python 1.6.1 2.415.7 1.0.0 epilot-template-variables"
     )
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None